Refactor(competianos): unificação de usecases, controllers e factories por funcionalidade
- antes controllers, usecases e factories eram separados em pastas diferentes de acordo com sua função (pastas controller, usecases e factories)
- agora são unificados por funcionalidade por completa dentro de competente, como utilizado dentro dos outros subdomínios
