+++ src/api/modules/competianos/usecases/deleteCompetiano/makeDeleteCompetianoUseCase.ts
@@ -1,5 +1,5 @@
 import { CompetianoMongoDBRepository } from "../../repositories/defaultMongoDBRepository/competianoRepository"
-import { DeleteCompetianoUseCase } from "../source/deleteCompetianoUseCase"
+import { DeleteCompetianoUseCase } from "./deleteCompetianoUseCase"
 
 export function makeDeleteCompetianoUseCase() {
     const usersRepository = new CompetianoMongoDBRepository()
