+++ src/api/modules/competianos/usecases/getCompetianoByEmail/getCompetianoByEmailUseCase.ts
